Move all styles to one file
