+++ client/src/Components/Checkbox/CheckboxComponent.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
-
 import { Checkbox } from './Checkbox';
-import { Label } from './Label';
+import { Label } from './CheckboxStyles';
 
 export const CheckboxComponent = () => {
   const [checked, setChecked] = useState(false);
