--- client/src/Components/Checkbox/CheckboxContainer.tsx
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const CheckboxContainer = styled.div`
-  display: flex;
-  align-items: center;
-`;
