--- client/src/Components/Checkbox/Icon.tsx
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const Icon = styled.svg`
-  fill: none;
-  stroke: white;
-  stroke-width: 2px;
-`;
