--- client/src/Components/Checkbox/Label.tsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-export const Label = styled.label`
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  width: 250px;
-  height: 35px;
-  background: #f2f4f3;
-  border-radius: 5px;
-`;
